Moved compute_metrics to a function (rather than method) to facilitate metrics at different levels of granularity.
